change mcbm table key type to uint32_t bc uint64_t causes masstree scan iterator bug

+++ benchmark/mcbm/table_decl.h
@@ -11,7 +11,7 @@
 
 namespace sql {
 
-FIELD(uint64_t, k);
+FIELD(uint32_t, k);
 KEYS(McbmKey);
